Code Refactoring
* Fix issue with 32bit architecture on big numbers' random generation

+++ p2p/strategy/nativeStrategy.go
@@ -305,13 +305,13 @@ func (ns *NativeStrategy) GetAnyResolvedPeer() *model.Peer {
 	if len(resolvedPeers) < 1 {
 		return nil
 	}
-	randomIdx := int(util.GetSecurePositiveRandom())
+	randomIdx := uint64(util.GetSecurePositiveRandom())
 	if randomIdx != 0 {
-		randomIdx %= len(resolvedPeers)
+		randomIdx %= uint64(len(resolvedPeers))
 	}
 	idx := 0
 	for _, peer := range resolvedPeers {
-		if idx == randomIdx {
+		if uint64(idx) == randomIdx {
 			return peer
 		}
 		idx++
@@ -365,10 +365,10 @@ func (ns *NativeStrategy) GetAnyUnresolvedPeer() *model.Peer {
 	if len(unresolvedPeers) < 1 {
 		return nil
 	}
-	randomIdx := int(util.GetSecurePositiveRandom()) % len(unresolvedPeers)
+	randomIdx := uint64(util.GetSecurePositiveRandom()) % uint64(len(unresolvedPeers))
 	idx := 0
 	for _, peer := range unresolvedPeers {
-		if idx == randomIdx {
+		if uint64(idx) == randomIdx {
 			return peer
 		}
 		idx++
@@ -493,10 +493,10 @@ func (ns *NativeStrategy) GetAnyKnownPeer() *model.Peer {
 	if len(knownPeers) < 1 {
 		panic("No well known peer is found")
 	}
-	randomIdx := int(util.GetSecurePositiveRandom()) % len(knownPeers)
+	randomIdx := uint64(util.GetSecurePositiveRandom()) % uint64(len(knownPeers))
 	idx := 0
 	for _, peer := range knownPeers {
-		if idx == randomIdx {
+		if uint64(idx) == randomIdx {
 			return peer
 		}
 		idx++
